Customize Bootstrap accent colors

Set accent colors (referred to as "brand" colors in the Bootstrap 3 docs, and "theme" colors in the Bootstrap 4 docs) for the current Bootstrap theme. Values must be NULL or a color string in a format htmltools::parseCssColors() can understand.

bs_theme_accent_colors(
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL
)

Arguments

primary
A color to be used for hyperlinks, to indicate primary/default actions, and to show active selection state in some Bootstrap components. Generally a bold, saturated color that contrasts with the theme's base colors.

secondary
A color for components and messages that don't need to stand out. (Not supported in Bootstrap 3.)

success
A color for messages that indicate an operation has succeeded. Typically green.

info
A color for messages that are informative but not critical. Typically a shade of blue-green.

warning
A color for warning messages. Typically yellow.

danger
A color for errors. Typically red.

See also

Examples

bs_theme_new("4+3")
bs_theme_accent_colors(primary = "maroon", secondary = "gray")
if (interactive()) {
  bs_theme_preview(with_themer = FALSE)
}