Add low-level theming customizations

Use bs_theme_add_variables() to set Bootstrap Sass variable defaults and bs_theme_add() to add arbitrary Sass code (i.e., styling rules, functions, mixins, etc).

bs_theme_add_variables(
  ...,
  .where = "defaults",
  .default_flag = identical(.where, "defaults")
)

bs_theme_add(defaults = "", declarations = "", rules = "", ...)

Arguments

...
For bs_theme_add_variables(), these arguments define Sass variables; otherwise, these arguments are passed along to sass::sass_layer().

.where
whether to place the variable definitions before other Sass "defaults", after other Sass "declarations", or after other Sass "rules".

.default_flag
whether or not to add a !default flag (if missing) to variable expressions. It's recommended to keep this as TRUE when .where = "defaults".

defaults
Any sass::as_sass() input to place before Bootstrap's Sass.

declarations
Any sass::as_sass() input to place after Bootstrap's variables, functions, and mixins, but before Bootstrap's styling rules.

rules
Any sass::as_sass() input to place after Bootstrap's Sass imports.

Details

Compared to higher-level theme customization functions (e.g., bs_theme_base_colors(), etc), these functions are a more direct interface to Bootstrap Sass, and therefore, do nothing to ensure the theme customization you add are portable between major Bootstrap versions.

References

https://getbootstrap.com/docs/4.4/getting-started/theming/

https://rstudio.github.io/sass/articles/sass.html#layering

See also

Examples

# Function to preview the styling a (primary) Bootstrap button
library(htmltools)
button <- tags$a(class = "btn btn-primary", href = "#", role = "button", "Hello")
preview_button <- function(x) {
  browsable(tags$body(x, button))
}

# To create a custom theme, you must start by calling bs_theme_new()
# Here we start with a theme based on a Bootswatch theme,
# then override some variable defaults
bs_theme_new(bootswatch = "sketchy")
bs_theme_add_variables(
  primary = "orange",
  "body-bg" = "#EEEEEE",
  "font-family-base" = "monospace",
  "font-size-base" = "1.4rem",
  "btn-padding-y" = ".16rem",
  "btn-padding-x" = "2rem",
  "border-radius" = 0,
  "border-radius-lg" = 0,
  "border-radius-sm" = 0
)
preview_button(bootstrap())

# If you need to set a variable based on another Bootstrap variable
bs_theme_add_variables(
  "body-color" = "$success",
  .where = "declarations"
)
preview_button(bootstrap())

# Start a new global theme and add some custom rules that
# use Bootstrap variables to define a custom styling for a
# 'person card'
bs_theme_new()
bs_theme_add(
  rules = sass::sass_file(
    system.file("custom", "person.scss", package = "bootstraplib")
  )
)
# Include custom CSS that leverages bootstrap Sass variables
person <- function(name, title, company) {
  tags$div(
    class = "person",
    h3(class = "name", name),
    div(class = "title", title),
    div(class = "company", company)
  )
}
browsable(tags$body(
  bootstrap(),
  person("Andrew Carnegie", "Owner", "Carnegie Steel Company"),
  person("John D. Rockefeller", "Chairman", "Standard Oil")
))