Customize Bootstrap typefaces

Set the typefaces used by Bootstrap for various purposes. Each argument can be NULL (no change), or a character vector of one or more elements.

bs_theme_fonts(base = NULL, code = NULL, heading = NULL)

Arguments

base
The default typeface.

code
The typeface to be used for code. Be sure this is monospace!

heading
The typeface to be used for heading elements.

Details

Each argument is a character vector, and each element of that vector can a single unquoted font family name, a single quoted font family name, or a comma-separated list of font families (with individual font family names quoted as necessary).

For example, each example below is valid:

# Single, unquoted
bs_theme_fonts(base = "Source Sans Pro")

# Single, quoted
bs_theme_fonts(base = "'Source Sans Pro'")

# Multiple, quoted
bs_theme_fonts(base = "'Source Sans Pro', sans-serif")

# Combining all of the above
bs_theme_fonts(base = c("Open Sans", "'Source Sans Pro'",
  "'Helvetica Neue', Helvetica, sans-serif"))

But the following is technically not valid:

# Incorrect--because multiple font families are being
# provided in a single string, names with spaces must
# be surrounded by quotes!
bs_theme_fonts(base = "Source Sans Pro, sans-serif")

The resulting CSS will contain font-family: Source Sans Pro, sans-serif; which is technically out of spec, but in fact is likely to still work with most browsers.

See also

Examples

bs_theme_new()
bs_theme_fonts(
  base = "Times",
  code = c("Courier", "monospace"),
  heading = "'Helvetica Neue', Helvetica, sans-serif"
)

if (interactive()) {
  bs_theme_preview(with_themer = FALSE)
}