Retrieve Sass variable values from the current theme

Useful for retriving a variable from the current theme and using the value to inform another R function.

bs_theme_get_variables(varnames)

Arguments

varnames
a character string referencing a Sass variable in the current theme.

Value

a character string containing a CSS/Sass value. If the variable(s) are not defined, their value is NA.

Examples

bs4_vars <- c("body-bg", "body-color", "primary", "border-radius")
bs_theme_new()
bs_theme_get_variables(bs4_vars)

#>       body-bg    body-color       primary border-radius 
#>        "#fff"     "#212529"     "#007bff"     "0.25rem" 

bs_theme_new(bootswatch = "darkly")
bs_theme_get_variables(bs4_vars)

#>       body-bg    body-color       primary border-radius 
#>        "#222"        "#fff"     "#375a7f"     "0.25rem"