Create a Bootstrap theme

bs_theme_new() creates a new (global) Bootstrap Sass theme which bootstrap() (or bootstrap_sass()) can consume (their theme argument defaults to bs_theme_get(), which get the current global theme). Once a global theme has been created, theme customization may be added via 'high-level' (e.g., bs_theme_base_colors()) and/or 'low-level' (e.g., bs_theme_add_variables()) customizations.

bs_theme_new(version = version_default(), bootswatch = NULL)

bs_theme_get()

bs_theme_clear()

bs_theme_set(theme)

Arguments

version
The major version of Bootstrap to use. A value of '4+3' means Bootstrap 4, but with additional CSS/JS to support BS3 style markup in BS4. Other supported versions include 3 and 4.

bootswatch
The name of a bootswatch theme. See bootswatch_themes() to list possible names.

theme
a theme object (i.e., the return value of bs_theme_get()).

References

https://getbootstrap.com/docs/4.4/getting-started/theming/

https://rstudio.github.io/sass/

See also

Examples

# Themes are stored globally, and if there is no theme present,
# bs_theme_get() returns NULL
bs_theme_get()

#> $defaults
#> /* Sass */
#> $primary: #008BA2 !default;
#> $navbar-height: 53 !default;
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/bs3compat/_defaults.scss";
#> $font-size-base: 0.875rem !default;
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_functions.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_variables.scss";
#> 
#> $declarations
#> /* Sass */
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_mixins.scss";
#> 
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/bs3compat/_declarations.scss";
#> 
#> 
#> 
#> $rules
#> /* Sass */
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_root.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_reboot.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_type.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_images.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_code.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_grid.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_tables.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_forms.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_buttons.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_transitions.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_dropdown.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_button-group.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_input-group.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_custom-forms.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_nav.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_navbar.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_card.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_breadcrumb.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_pagination.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_badge.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_jumbotron.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_alert.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_progress.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_media.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_list-group.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_close.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_toasts.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_modal.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_tooltip.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_popover.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_carousel.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_spinners.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_utilities.scss";
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap/scss/_print.scss";
#> .table th[align=left] { text-align: left; }
#> .table th[align=right] { text-align: right; }
#> .table th[align=center] { text-align: center; }
#> @import "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/bs3compat/_rules.scss";
#> 
#> 
#> 
#> $html_deps
#> $html_deps[[1]]
#> List of 10
#>  $ name      : chr "bs3compat"
#>  $ version   : chr "0.1.0.9000"
#>  $ src       :List of 1
#>   ..$ file: chr "bs3compat/js"
#>  $ meta      : NULL
#>  $ script    : chr [1:2] "tabs.js" "bs3compat.js"
#>  $ stylesheet: NULL
#>  $ head      : NULL
#>  $ attachment: NULL
#>  $ package   : chr "bootstraplib"
#>  $ all_files : logi TRUE
#>  - attr(*, "class")= chr "html_dependency"
#> 
#> 
#> $file_attachments
#>                                                                                                                                        fonts 
#> "/private/var/folders/t8/jxw4bftj35g15fcb6m4mfbc80000gn/T/Rtmp06qkXD/temp_libpath12f24607b0bf2/bootstraplib/lib/bootstrap-sass/assets/fonts" 
#> 
#> $tags
#> [1] "bootstraplib_version_4+3"
#> 
#> attr(,"class")
#> [1] "bs_theme"   "sass_layer"

# Use latest Bootstrap version, default theme
bs_theme_new()

# When a theme is set, bs_theme_get() returns all the Sass code
# necessary to compile Bootstrap, plus any theming customizations.
class(bs_theme_get())

#> [1] "bs_theme"   "sass_layer"

# If you just want to know the
theme_version()

#> [1] "4+3"

# Preview the current theme styling in a shiny app
if (interactive()) bs_theme_preview()

# Latest Bootstrap version, minty theme
bs_theme_new(bootswatch = "minty")
theme_bootswatch()

#> [1] "minty"

if (interactive()) bs_theme_preview()

# Use Bootstrap 3 with theme customization
bs_theme_new(3)
bs_theme_base_colors(bg = "#444", fg = "#e4e4e4")
bs_theme_accent_colors(primary = "#e39777")
if (interactive()) bs_theme_preview(with_themer = FALSE)

# Since themes are sass layers, you can work with them
# locally as if they were any other Sass layer
# (i.e. you don't have to modify global state to add theme customizations)
bs_theme_new("4+3")
theme <- sass::sass_layer_merge(
  bs_theme_get(),
  sass::sass_layer(defaults = list("primary" = "red !default"))
)
bootstrap_sass(
  ".my-class { color: mix($primary, $secondary, 50%); }",
  theme
)

#> /* CSS */
#> .my-class {
#>   color: #b63b3f;
#> }
#>